requirements/mnms.txt
django-bakery
